requirements.txt
Flask==0.12.2
Keras==2.1.2
numpy==1.11.0
Pillow==5.1.0
h5py